examples/django_mongoengine/requirements.txt
Django==2.2.3
pytest==4.6.3
pytest-django==3.5.1
mongoengine==0.17.0
mongomock==3.16.0
graphene-django==2.4.0
graphene-mongo
